add react navigation and screens
