+++ package.json
@@ -10,12 +10,19 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.7",
+    "@react-navigation/native": "^5.0.10",
+    "@react-navigation/stack": "^5.2.0",
     "emoji-datasource": "^5.0.1",
     "emoji-datasource-apple": "^5.0.1",
     "emoji-datasource-google": "^5.0.1",
     "react": "16.9.0",
     "react-native": "0.61.5",
-    "react-native-fast-image": "^8.1.5"
+    "react-native-fast-image": "^8.1.5",
+    "react-native-gesture-handler": "^1.6.0",
+    "react-native-reanimated": "^1.7.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.7",
